QUnit: move return value from *int back to int
Explanation: the main routine _must_ return an integer value to make
sure the correct return status is propagated to the caller, eg.
`./testScript.qtest && echo yes' must correctly handle the return value
of the test script. For that reason, the proposed %exec-class for which
this change was necessary cannot be used.

git-svn-id: svn+ssh://svn.code.sf.net/p/qore/code/qore/trunk@7229 1b0b35fe-f740-4c1b-b404-a338b32e513b