Fix MySQL datatype definitions for [var]char and [var]binary
The CHAR and VARCHAR types are declared with a length that indicates the maximum number of characters you want to store.
The length of a CHAR column can be any value from 0 to 255.
The length of a VARCHAR column can be specified as a value from 0 to 65,535.

The permissible maximum length is the same for BINARY and VARBINARY as it is for CHAR and VARCHAR, except that the length for BINARY and VARBINARY is a length in bytes rather than in characters.