Cleanup cmake build.
Added the pseudo qpp files to a separate variable.
Added missing cpp files to the libqore cpp sources.
Removed the qpp generated files from the libqore cpp sources.
Added the variable holding the qpp generated cpp sources to add_library of
libqore.
Fixed typo in QoreMacrosIntern.cmake.

Due to that the pseudo qpp sources are now in a separate variable and not
among the other qpp files, this build now works without having the qpp
genarated files manually in the cpp source list.