refs #4798 Fix build on illumos
refs #4798 Fix build on illumos.

Illumos defines both AF_LINK and AF_PACKET
but does not have link_ntoa.
Use the finding of net/if_dl.h and link_ntoa
as indicators instead of AF_LINK.
AF_PACKET is supported on illumos but it
does not have linux/if_packet.h and it is
not needed to read AF_PACKET from getifaddrs.
So use configure checks to see if it is to
be included.

The linker on illumos does not understand
"-undefined dynamic_lookup". It seems to be
macosx specific but tolerated by GNU ld.
So only use it when building on a mac.